+++ spec/nanook/node_spec.rb
@@ -171,6 +171,54 @@ RSpec.describe Nanook::Node do
     expect(Nanook.new.node.sync_progress).to eq 99.50248756218906
   end
 
+  it "should show synchronizing_blocks" do
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"unchecked\",\"count\":\"1000\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"blocks\":{\"000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F\":\"{\\\"type\\\": \\\"open\\\",\\\"account\\\": \\\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\\\",\\\"representative\\\": \\\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\\\",\\\"source\\\": \\\"FA5B51D063BADDF345EFD7EF0D3C5FB115C85B1EF4CDE89D8B7DF3EAF60A04A4\\\",\\\"work\\\": \\\"0000000000000000\\\",\\\"signature\\\":\\\"00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\\\"}\",\"000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3C\":\"{\\\"type\\\": \\\"open\\\",\\\"account\\\": \\\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\\\",\\\"representative\\\": \\\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\\\",\\\"source\\\": \\\"FA5B51D063BADDF345EFD7EF0D3C5FB115C85B1EF4CDE89D8B7DF3EAF60A04A4\\\",\\\"work\\\": \\\"0000000000000000\\\",\\\"signature\\\":\\\"00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\\\"}\"}}",
+      headers: {}
+    )
+
+    response = Nanook.new.node.synchronizing_blocks
+
+    expect(response).to have(2).items
+
+    block = response["000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F"]
+
+    expect(block[:type]).to eq "open"
+    expect(block[:account]).to eq "xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000"
+    expect(block[:representative]).to eq "xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000"
+    expect(block[:source]).to eq "FA5B51D063BADDF345EFD7EF0D3C5FB115C85B1EF4CDE89D8B7DF3EAF60A04A4"
+    expect(block[:work]).to eq "0000000000000000"
+    expect(block[:signature]).to eq "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+  end
+
+  it "should show synchronizing_blocks with limit" do
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"unchecked\",\"count\":\"1\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"blocks\": {\"000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F\": \"{\\\"type\\\": \\\"open\\\",\\\"account\\\": \\\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\\\",\\\"representative\\\": \\\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\\\",\\\"source\\\": \\\"FA5B51D063BADDF345EFD7EF0D3C5FB115C85B1EF4CDE89D8B7DF3EAF60A04A4\\\",\\\"work\\\": \\\"0000000000000000\\\",\\\"signature\\\":\\\"00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\\\"}\"}}",
+      headers: {}
+    )
+
+    response = Nanook.new.node.synchronizing_blocks(limit: 1)
+
+    expect(response).to have(1).item
+
+    block = response["000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F"]
+
+    expect(block[:type]).to eq "open"
+    expect(block[:account]).to eq "xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000"
+    expect(block[:representative]).to eq "xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000"
+    expect(block[:source]).to eq "FA5B51D063BADDF345EFD7EF0D3C5FB115C85B1EF4CDE89D8B7DF3EAF60A04A4"
+    expect(block[:work]).to eq "0000000000000000"
+    expect(block[:signature]).to eq "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+  end
+
   it "should synced? when true" do
     stub_request(:post, uri).with(
       body: "{\"action\":\"block_count\"}",